fresh-ui: repair the suite after the render-object rewrite
Window-dependent layout readers now invalidate when the enclosing scroll
window moves: the constraint cache is keyed by constraints, and a window
is not one, so a reader that reads the window returned its last answer
after an Anchor::reveal. RenderObject gains reads_window(), cached on the
render node, and set_scroll walks descendants to mark those nodes dirty,
stopping at nested clipping nodes.

The stacked-hit-path test now asserts that transparent handlers on a
lower path run, which is the behaviour R7 introduced; it previously
recorded the weaker single-path behaviour. Two goldens are updated: both
recorded stale-frame bugs (a deleted row lingering, a theme change
arriving one frame late) that the render tree fixes.

+++ crates/fresh-ui/src/test/screen.rs
@@ -115,8 +115,8 @@ fn draw(s: &mut Screen, item: &Item, frame: Rect, fill_char: &impl Fn(&str) -> O
             window,
         } => {
             let track = r.h.max(1);
-            let thumb = ((*window as u32 * track as u32) / (*content).max(1) as u32).max(1) as u16;
-            let top = ((*offset as u32 * track as u32) / (*content).max(1) as u32) as u16;
+            let thumb = ((*window as u32 * track as u32) / (*content).max(1)).max(1) as u16;
+            let top = ((*offset * track as u32) / (*content).max(1)) as u16;
             for i in 0..track {
                 let ch = if i >= top && i < top + thumb {
                     '█'
